requirements.txt
aiogram==2.25.1
alembic==1.10.2
Flask==2.2.3
Flask_Login==0.6.2
Flask_Migrate==4.0.4
flask_sqlalchemy==3.0.3
Flask_WTF==1.1.1
mysql_connector_repackaged==0.3.1
SQLAlchemy==2.0.7
Werkzeug==2.3.4
WTForms==3.0.1
mysqlclient==2.1.1
email-validator==2.0.0.post2
mysql-connector==2.2.9
mysql-connector-python~=8.0.33
Pillow~=9.4.0